move common logic, agent/common/svid -> common/rotationutil
